+++ com.unity.cinemachine/Runtime/Deprecated/Cinemachine3rdPersonFollow.cs
@@ -201,9 +201,7 @@ namespace Cinemachine
         {
             base.OnTargetObjectWarped(target, positionDelta);
             if (target == FollowTarget)
-            {
                 m_PreviousFollowTargetPosition += positionDelta;
-            }
         }
         
         void PositionCamera(ref CameraState curState, float deltaTime)