+++ com.unity.cinemachine/Runtime/PostProcessing/CinemachinePostProcessing.cs
@@ -86,11 +86,7 @@ namespace Cinemachine
         [FormerlySerializedAs("m_Profile")]
         public PostProcessProfile Profile;
 
-        /// <summary>Legacy support for obsolete format</summary>
-        [HideInInspector, SerializeField, FormerlySerializedAs("m_FocusTracksTarget")]
-        bool m_LegacyFocusTracksTarget;
-       
-        class VcamExtraState
+        class VcamExtraState : VcamExtraStateBase
         {
             public PostProcessProfile ProfileCopy;
 
@@ -117,16 +113,10 @@ namespace Cinemachine
             }
         }
 
+        List<VcamExtraState> m_extraStateCache;
+
         /// <summary>True if the profile is enabled and nontrivial</summary>
         public bool IsValid => Profile != null && Profile.settings.Count > 0;
-
-        /// <summary>Called by the editor when the shared asset has been edited</summary>
-        public void InvalidateCachedProfile()
-        {
-            var list = GetAllExtraStates<VcamExtraState>();
-            for (int i = 0; i < list.Count; ++i)
-                list[i].DestroyProfileCopy();
-        }
 
         void OnValidate()
         {
@@ -144,21 +134,22 @@ namespace Cinemachine
 
         protected override void OnEnable()
         {
-            base.OnEnable();
-
-            // Map legacy m_FocusTracksTarget to focus mode
-            if (m_LegacyFocusTracksTarget)
-            {
-                FocusTracking = VirtualCamera.LookAt != null 
-                    ? FocusTrackingMode.LookAtTarget : FocusTrackingMode.Camera;
-            }
-            m_LegacyFocusTracksTarget = false;
+            InvalidateCachedProfile();
         }
 
         protected override void OnDestroy()
         {
             InvalidateCachedProfile();
             base.OnDestroy();
+        }
+
+        /// <summary>Called by the editor when the shared asset has been edited</summary>
+        public void InvalidateCachedProfile()
+        {
+            m_extraStateCache ??= new();
+            GetAllExtraStates(m_extraStateCache);
+            for (int i = 0; i < m_extraStateCache.Count; ++i)
+                m_extraStateCache[i].DestroyProfileCopy();
         }
 
         /// <summary>Apply PostProcessing effects</summary>
@@ -200,7 +191,7 @@ namespace Cinemachine
                                 switch (FocusTracking)
                                 {
                                     default: break;
-                                    case FocusTrackingMode.FollowTarget: focusTarget = VirtualCamera.Follow; break;
+                                    case FocusTrackingMode.FollowTarget: focusTarget = vcam.Follow; break;
                                     case FocusTrackingMode.CustomTarget: focusTarget = FocusTarget; break;
                                 }
                                 if (focusTarget != null)